move tables and tree into ls
